refactor: merge with existing controllers
