+++ src/main/java/ch/presentium/backend/api/user/request/StudentRequestBody.java
@@ -1,0 +1,10 @@
+package ch.presentium.backend.api.user.request;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
+public record StudentRequestBody(
+    @NotBlank String firstName,
+    @NotBlank String lastName,
+    @NotBlank @Email String email
+) {}
