--- src/main/java/ch/presentium/backend/business/service/StudentService.java
@@ -1,36 +1,0 @@
-package ch.presentium.backend.business.service;
-
-import ch.presentium.backend.business.model.user.Student;
-import ch.presentium.backend.business.repository.StudentRepository;
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class StudentService {
-
-    private final StudentRepository studentRepository;
-
-    @Autowired
-    public StudentService(StudentRepository studentRepository) {
-        this.studentRepository = studentRepository;
-    }
-
-    public List<Student> findAll() {
-        return studentRepository.findAll();
-    }
-
-    public Optional<Student> findById(UUID id) {
-        return studentRepository.findById(id);
-    }
-
-    public void addStudent(Student student) {
-        studentRepository.save(student);
-    }
-
-    public void deleteStudent(UUID id) {
-        studentRepository.deleteById(id);
-    }
-}
